atualizacao da listagem

+++ app/src/main/java/br/com/allanlarangeiras/desafioorama/model/dto/Operability.kt
@@ -5,5 +5,5 @@ import java.io.Serializable
 
 class Operability(
     @SerializedName("minimum_initial_application_amount")
-    val minimumInitialApplicationAmount: Int
+    val minimumInitialApplicationAmount: Double
 ): Serializable
